+++ Source/Mosa.Compiler.Framework/Transform/Manual/IR/LowerTo32/SignExtend16x64.cs
@@ -24,7 +24,7 @@ namespace Mosa.Compiler.Framework.Transform.Manual.IR.LowerTo32
 
 			context.SetInstruction(IRInstruction.Move32, v1, operand1);
 			context.AppendInstruction(IRInstruction.SignExtend16x32, resultLow, v1);
-			context.AppendInstruction(IRInstruction.ArithShiftRight32, resultHigh, resultLow, transformContext.CreateConstant((byte)31));
+			context.AppendInstruction(IRInstruction.ArithShiftRight32, resultHigh, resultLow, transformContext.CreateConstant(31));
 			context.AppendInstruction(IRInstruction.To64, result, resultLow, resultHigh);
 		}
 	}
